+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Use `later::global_loop()` instead of `later::current_loop()` in `controller$autoscale()`.
 * Test `crew_worker()` with `autoexit = FALSE` (https://github.com/r-lib/mirai/issues/535).
+* Deprecate the `dispatchers()` method of `crew_monitor_local()` objects because the dispatcher is a thread in `mirai` > 2.6.1.
 
 # crew 1.3.0